Migrate schema URN tests.
Migrated the schema tests to be placed under the
GenericScimResourceObjectTest and new BaseScimResourceTest classes.
